Update pollyxt.py
line 167: "...!= 0", the 0 should be a variable (e.g. depol_angle_regular_measurement) which is read from locations.ini. for some polly systems it is a different angle.
E.g: 
Arielle: 19.98755
TROPOS: 999
NOA: 000000 
1v2: 0.0000
tau: 350
tjk: 0000

Furthermore, it is important to skip the first and last measurement of the period when the depol_cal_angle is not regular. Because the filter wheel is storing its position only once during the 30 seconds measurement. So be on the save side, you should skip the first and last of the +45 period, and the first and last of the - 45 period. 

Example: PollyXT_FMI:

index: 
420	421	422	423	424	425	426	427	428	429	430	431	432	433	434	435	436	437	438	439	440	441	442	443
Depol cal angle:
0.0	151	151	151	151	151	151	151	151	151	151	241	241	241	241	241	241	241	241	241	241	255	0.0	0.0
--> use for +45: 422--429 and for -45 432--439

Thank you very much,this will help the improve the Polly SCC performance a lot!